Refactor of ViewSelect
