--- packages/sentry-client/src/sentry/constants/serverView.ts
@@ -1,9 +1,0 @@
-interface IServerView {
-    [key: string]: string
-}
-
-export const ServerView: IServerView = {
-    servers: "Servers",
-    hybrid: "Hybrid",
-    services: "Services"
-}
