+++ packages/sentry-client/src/sentry/constants/view.ts
@@ -1,0 +1,10 @@
+interface IView {
+    [key: string]: string;
+}
+
+const view: IView = {
+    servers: "Servers",
+    services: "Services"
+}
+
+export default view;
